Add support for UI extensions
In preparation of #34, this change introduces a mechanism to register UI
extensions. UI extensions allow to show additional UI elements (HTML
elements) on top of Sprotty diagrams. Such UI extensions can be
registered with an ID and then enabled or disabled by dispatching
respective actions for showing and hiding extensions.

A UI extension is rendered in a separate div element but can be
positioned on top of Sprotty diagrams based on a current context (e.g.
selected elements, etc.).

Signed-off-by: Philip Langer <planger@eclipsesource.com>